Removed bad cast to float in IntImm
Former-commit-id: 6147be1f40165e5a6a2c9fec0542c16b34767b64

+++ cpp/IR.h
@@ -36,7 +36,7 @@ namespace HalideInternal {
 
     struct IntImm : public Expr {
         int value;
-        IntImm(float v) : value(v) {}
+        IntImm(int v) : value(v) {}
 
         void accept(IRVisitor *v) const {v->visit(this);}
     };
